requirements.txt
django>=4.2
djangorestframework
django-cors-headers
pymupdf
python-docx
anthropic
python-dotenv